Working around a weird liferay bug regarding shared classes
When dealing with classes shared by several portlets, in Liferay there are usually the following ways to resolve this: create a libary JAR with all those classes and a) copy them to TOMCAT/lib/ext, b) copy them to TOMCAT/webapps/ROOT/WEB-INF/lib and add the JAR as required in each portlet's liferay-plugin-package.properties, c) include the JAR in each and every portlet - additionally specify deployment-dependencies in liferay-plugin-package.properties. However, this time we got a ClastCastException: X cannot be cast to X in all three scenarios. Why this happens (only for SOME classes of the shared JAR) is not yet clear, but it is really weird. Currently workarounding this by invoking service-builder methods without custom objects as parameters.